+++ NEWS.md
@@ -4,9 +4,22 @@ piecepackr.metadata (development)
 Initial features
 ----------------
 
+* `docinfo()` provides an `{R6}` object to represent pdf documentation info entries.
+* `get_docinfo()` reads in pdf documentation info entries from a pdf file.
+  `get_docinfo()` will try to use the following helper functions in the following order:
+
+  1. `get_docinfo_pdftools()` which wraps `pdftools::pdf_info()`
+  2. `get_docinfo_pdftk()` which wraps `pdftk` command-line tool
+
+* `set_docinfo()` sets (replaces) pdf documentation info entries to a pdf file (#3).
+  `set_docinfo()` will try to use the following helper functions in the following order:
+
+  1. `set_docinfo_gs()` which wraps `ghostscript` command-line tool
+  2. `set_docinfo_pdftk()` which wraps `pdftk` command-line tool
+
 * `n_pages()` returns the number of pages in the (pdf) file(s).
   `n_pages()` will try to use the following helper functions in the following order:
  
-  1. `n_pages_qpdf()` which wraps `qpdf::pdf_length()`.
+  1. `n_pages_qpdf()` which wraps `qpdf::pdf_length()`
   2. `n_pages_pdftk()` which wraps `pdftk` command-line tool
   3. `n_pages_gs()` which wraps `ghostscript` command-line tool